shinysnap: Save and Restore the State of 'shiny' Applications

Save the state of applications built with 'shiny', the web application framework by Chang et al. (2026) doi:10.32614/CRAN.package.shiny . Users can share their work and continue it in another session. Input values and selected values held by the server are saved in JSON (JavaScript Object Notation) files that can be read and edited by hand. Saved state can be restored without reloading the page or setting up bookmarking. Restoration waits for inputs that appear as the page changes and reports which values were restored, missing, or could not be applied.

See also

Author

Maintainer: Nan Xiao me@nanx.me (ORCID) [copyright holder]

Authors:

Nan Xiao me@nanx.me (ORCID) [copyright holder]